Fixes, hopefully all working now MUST TEST YET

+++ src/code/RemovedNotice.java
@@ -13,16 +13,31 @@ public class RemovedNotice implements Runnable {
         this.fileId = fileId;
         this.chunkNo = chunkNo;
 
-        File chunk = new File("peer" + Peer.senderId + "/backup/" + fileId + "/chk" + chunkNo);
-        if(!chunk.delete()){
-            System.out.println("Error deleting file");
-            System.exit(-1);
+        File dir = new File("peer" + Peer.senderId + "/backup/" + fileId);
+
+        if(dir.exists() && dir.isDirectory()){
+            File chunk = new File("peer" + Peer.senderId + "/backup/" + fileId + "/chk" + chunkNo);
+            if(!chunk.delete()){
+                System.out.println("Error deleting file " + fileId + " #" + chunkNo);
+                System.exit(-1);
+            }
+
+            String[] list = dir.list();
+
+            if(list != null && list.length == 0){
+                if(!dir.delete()){
+                    System.out.println("Error deleting dir " + fileId);
+                    System.exit(-1);
+                }
+
+                System.out.println("Deleted dir " + fileId);
+            }
+
+            System.out.println("Deleted file " + fileId + " #" + chunkNo + " - Used Space: " + Peer.getUsedSpace());
+
+            Key key = new Key(fileId, chunkNo);
+            Peer.stores.get(key).decrement();
         }
-
-        System.out.println("Deleted file " + fileId + " #" + chunkNo + " - Used Space: " + Peer.getUsedSpace());
-
-        Key key = new Key(fileId, chunkNo);
-        Peer.stores.get(key).decrement();
     }
 
     @Override
